requirements.txt
FPDF
python-dotenv==0.19.2
PyPDF2==3.0.0
reportlab==3.6.12